Added code for reordering the play queue

+++ app/src/main/java/com/codersguidebook/music/PlayQueueViewModel.kt
@@ -8,7 +8,7 @@ import androidx.lifecycle.ViewModel
 
 class PlayQueueViewModel : ViewModel() {
     var playQueue = MutableLiveData<List<QueueItem>>()
-    var currentQueueItemId = MutableLiveData<Int>()
+    var currentQueueItemId = MutableLiveData<Long>()
     var currentlyPlayingSongMetadata = MutableLiveData<MediaMetadataCompat?>()
     var playbackDuration = MutableLiveData<Int>()
     var playbackPosition = MutableLiveData<Int>()
